requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==5.2.4
django-graphql-jwt==0.4.0
exceptiongroup==1.3.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kombu==5.5.4
openai==1.97.1
packaging==25.0
promise==2.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
redis==6.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
django_celery_results==2.6.0
django-filter==25.1
python-decouple>=3.6
psycopg2-binary==2.9.10
google-auth==2.40.3
django-graphql-jwt==0.4.0
requests==2.32.4
graphene-file-upload==1.3.0
Pillow==10.0.0
pytesseract==0.3.10
